requirements.txt
Preferences==0.08
python-xlib=0.25
